fix periods

+++ src/arduino/smart_cm/Config.h
@@ -23,9 +23,9 @@
 const float waterLevel_1 = 33;
 const float waterLevel_2 = 66;
 const float waterLevel_max = 100;
-const int period_normal = 3000;
-const int period_preAlarm = 2000;
-const int period_alarm = 1000;
+const long period_normal = 3000;
+const long period_preAlarm = 2000;
+const long period_alarm = 1000;
 
 const int lightTreshold = 0;
 const int lightTimer = 0;
